Safari-609.1.11.1/LayoutTests/crypto/workers/crypto-gc-worker-expected.txt

[Worker] Test that self.crypto wrapper preserves custom properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/crypto-gc-worker.js
PASS [Worker] self.crypto.foo is "bar"
PASS successfullyParsed is true

TEST COMPLETE
